classic/requirements.txt
-c constraints.txt
-e src/plonedemo.site[test]

zope.testrunner

# Add required add-ons
# (Ideally add them in setup.py for plone6demo)
# collective.easyform
